services/requirements.txt
playwright==1.44.0
fastapi==0.111.0
uvicorn[standard]>=0.31.1
sqlalchemy==2.0.31
browser-use==0.12.6
python-dotenv==1.2.1
langfuse==4.3.1
openinference-instrumentation-openai==0.1.44
